Refactor to separate data and presentation, reorder folders structure, remove unnecessary files and folders
