+++ src/App.jsx
@@ -3,10 +3,10 @@ import { useSelector } from "react-redux"
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom"
 import { Footer } from "./components/footer/Footer"
 import { Header } from "./components/header/Header"
-import { Home } from "./components/Home"
-import { Account } from "./components/Account"
-import { Login } from "./components/Login"
-import { Register } from "./components/Register"
+import { Home } from "./pages/Home"
+import { Account } from "./pages/Account"
+import { Login } from "./pages/Login"
+import { Register } from "./pages/Register"
 
 const App = () => {
   const authenticated = useSelector((state) => state.auth.authenticated)
