+++ src/assets/data/order.js
@@ -1,0 +1,22 @@
+export const order = [
+  {
+    id: 1,
+    title: "Your Order",
+    desc: "There are many variations of passages of Lorem Ipsum available",
+  },
+  {
+    id: 2,
+    title: "Ready Your Order",
+    desc: "There are many variations of passages of Lorem Ipsum available",
+  },
+  {
+    id: 3,
+    title: "Packing Your Order",
+    desc: "There are many variations of passages of Lorem Ipsum available",
+  },
+  {
+    id: 4,
+    title: "And Deliver",
+    desc: "There are many variations of passages of Lorem Ipsum available",
+  },
+]
