+++ src/components/UI/ProductCart.jsx
@@ -4,15 +4,15 @@ import { useDispatch } from "react-redux"
 import { Link } from "react-router-dom"
 import { cartActions } from "../../store"
 
-export const ProductCart = ({ key, id, cover, name, price }) => {
+export const ProductCart = ({ id, cover, name, price }) => {
   const dispatch = useDispatch()
   const addToCart = () => {
     dispatch(cartActions.addToCart({ id, name, price, cover }))
   }
   return (
-    <div className="box boxItems" id="product" onClick={addToCart}>
+    <div className="box boxItems" id="product" onClick={addToCart} key={id}>
       <div className="img">
-        <Link>
+        <Link to="/#">
           <img src={cover} alt="cover" />
         </Link>
       </div>
