+++ src/components/hero/Order.jsx
@@ -1,28 +1,7 @@
 import React from "react"
 import "./slider.css"
+import { order } from '../../assets/data/order';
 
-const order = [
-  {
-    id: 1,
-    title: "Your Order",
-    desc: "There are many variations of passages of Lorem Ipsum available",
-  },
-  {
-    id: 2,
-    title: "Ready Your Order",
-    desc: "There are many variations of passages of Lorem Ipsum available",
-  },
-  {
-    id: 3,
-    title: "Packing Your Order",
-    desc: "There are many variations of passages of Lorem Ipsum available",
-  },
-  {
-    id: 4,
-    title: "And Deliver",
-    desc: "There are many variations of passages of Lorem Ipsum available",
-  },
-]
 export const category = [
   {
     id: 1,
