+++ src/pages/DetailsPages.jsx
