+++ src/pages/Register.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import "./login.css"
+import "../styles/login.css"
 
 export const Register = () => {
   return (
