+++ src/styles/details.css
